+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to `laravel-newsletter` will be documented in this file.
 
+## 5.0.0 - 2022-11-10
+
+- the package is now driver based
+- added driver from [Mailcoach](https://mailcoach.app) and [MailChimp](https://mailchimp.com)
+- simplified code
+- refactored tests to Pest
+
 ## 4.11.0 - 2022-01-14
 
 - allow Laravel 9